本地端/requirements.txt
PyQt5
openpyxl
qrcode
Pillow
pandas
numpy
Flask
requests